refactor: move semantic and settings services to correct layers
Addresses issue #242 by relocating services from Api to the layers that
own the corresponding concerns, enforcing the Application/Infrastructure/Api
dependency hierarchy.

Moves:
- StubSettings + SemanticMatchingSettings: Infrastructure → Application
  (application-level config POCOs; no Infrastructure-only behavior)
- Semantic/* (ISemanticMatcherService, SemanticMatcherService, helpers):
  Api → Application (pure domain logic; no ASP.NET dependency)
- ISemanticEmbeddingClient: Api → Application (interface belongs with
  its consumer; Infrastructure provides the implementation)
- SemanticEmbeddingClient: Api → Infrastructure (HTTP I/O belongs in
  Infrastructure; implements the Application interface)
- StubResponseHeaderBuilder: Api → Application (pure header conversion
  utility; no Api-model or ASP.NET dependency)

SemanticMatcherService constructor now accepts StubSettings directly
instead of IOptions<StubSettings> so Application stays free of
Microsoft.Extensions.Options. Call sites in Api pass .Value.

Not moved (require cascading model migration — follow-up PR):
- Resolution orchestration (StubService, StubDispatchSelector, etc.)
- Inspection services
- Matching/StubDefaultResponseSelector
These all depend on StubResponse/StubMatchResult (Api.Models) or
HttpMethods (ASP.NET Core), which cannot yet be referenced from Application.
