+++ src/SemanticStub.Api/Services/Resolution/StubDispatchSelector.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Primitives;
 using SemanticStub.Api.Models;
 using SemanticStub.Application.Models;
 using SemanticStub.Application.Services;
+using SemanticStub.Application.Services.Semantic;
 
 namespace SemanticStub.Api.Services;
