+++ src/SemanticStub.Application/Infrastructure/Yaml/StubSettings.cs
@@ -1,4 +1,4 @@
-namespace SemanticStub.Infrastructure.Yaml;
+namespace SemanticStub.Application.Infrastructure.Yaml;
 
 /// <summary>
 /// Configures the YAML stub definition source and optional matching features.
