+++ src/SemanticStub.Application/Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SemanticStub.Application.Tests")]
 [assembly: InternalsVisibleTo("SemanticStub.Api.Tests")]
+[assembly: InternalsVisibleTo("SemanticStub.Api")]
