+++ src/SemanticStub.Application/Services/Semantic/SemanticEmbeddingEndpoint.cs
@@ -1,8 +1,8 @@
-namespace SemanticStub.Api.Services;
+namespace SemanticStub.Application.Services.Semantic;
 
-internal static class SemanticEmbeddingEndpoint
+public static class SemanticEmbeddingEndpoint
 {
-    internal static string Normalize(string endpoint)
+    public static string Normalize(string endpoint)
     {
         var normalized = endpoint.TrimEnd('/');
         return normalized.EndsWith("/embed", StringComparison.OrdinalIgnoreCase)
