+++ src/SemanticStub.Infrastructure/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("SemanticStub.Api.Tests")]
+[assembly: InternalsVisibleTo("SemanticStub.Api")]
