+++ tests/SemanticStub.Api.Tests/Unit/Resolution/StubResponseHeaderBuilderTests.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Primitives;
 using SemanticStub.Api.Models;
 using SemanticStub.Application.Models;
-using SemanticStub.Api.Services;
+using SemanticStub.Application.Services.Resolution;
 using Xunit;
 
 namespace SemanticStub.Api.Tests.Unit.Resolution;
