+++ tests/SemanticStub.Api.Tests/Unit/Resolution/StubServiceTests.cs
@@ -7,6 +7,7 @@ using SemanticStub.Infrastructure.Yaml;
 using SemanticStub.Api.Models;
 using SemanticStub.Application.Models;
 using SemanticStub.Application.Services;
+using SemanticStub.Application.Services.Semantic;
 using SemanticStub.Api.Services;
 using Xunit;
